+++ Assets/ML-Agents/Timers/step1_timers.json
@@ -1,1 +1,1 @@
-{"count":1,"self":299.90356479999997,"total":300.0273876,"children":{"InitializeActuators":{"count":1,"self":0.0019944999999999997,"total":0.0019944999999999997,"children":null},"InitializeSensors":{"count":1,"self":0.0019939,"total":0.0019939,"children":null},"AgentSendState":{"count":1696,"self":0.0165446,"total":0.0516791,"children":{"CollectObservations":{"count":340,"self":0.0064912,"total":0.0064912,"children":null},"WriteActionMask":{"count":340,"self":0,"total":0,"children":null},"RequestDecision":{"count":340,"self":0.0058924,"total":0.0286433,"children":{"RayPerceptionSensor.Perceive":{"count":340,"self":0.022750899999999998,"total":0.022750899999999998,"children":null}}}}},"DecideAction":{"count":1696,"self":0.042557899999999996,"total":0.042557899999999996,"children":null},"AgentAct":{"count":1696,"self":0.025602399999999997,"total":0.025602399999999997,"children":null},"RayPerceptionSensor.Perceive":{"count":2,"self":0,"total":0,"children":null}},"gauges":{"MLTank.CumulativeReward":{"count":2,"max":-1,"min":-1,"runningAverage":-1,"value":-1,"weightedAverage":-1}},"metadata":{"timer_format_version":"0.1.0","start_time_seconds":"1614509243","unity_version":"2019.4.18f1","command_line_arguments":"D:\\Unity\\2019.4.18f1\\Editor\\Unity.exe -projectpath D:\/Project\/UNITY\/ml-agents\/ML-tanks -useHub -hubIPC -cloudEnvironment production -hubSessionId 6926a880-795e-11eb-8b20-5718f046107e -accessToken LdsvsevxA4tdhBjYCUDmE5IPvB4sHNv_rwWWDJW76us00ef","communication_protocol_version":"1.3.0","com.unity.ml-agents_version":"1.7.2-preview","scene_name":"step1","end_time_seconds":"1614509543"}}
+{"count":1,"self":280.2902784,"total":864.5606159,"children":{"InitializeActuators":{"count":1,"self":0.0019755,"total":0.0019755,"children":null},"InitializeSensors":{"count":1,"self":0.0019946,"total":0.0019946,"children":null},"AgentSendState":{"count":315267,"self":1.8172519999999999,"total":9.863795099999999,"children":{"CollectObservations":{"count":63054,"self":0.37958739999999996,"total":0.37958739999999996,"children":null},"WriteActionMask":{"count":63054,"self":0.1654481,"total":0.1654481,"children":null},"RequestDecision":{"count":63054,"self":0.96239619999999992,"total":7.5015076999999994,"children":{"AgentInfo.ToProto":{"count":63054,"self":0.5066499,"total":6.5391115,"children":{"GenerateSensorData":{"count":63054,"self":1.3377788,"total":6.0324615999999995,"children":{"RayPerceptionSensor.Perceive":{"count":63054,"self":4.6946828,"total":4.6946828,"children":null}}}}}}}}},"DecideAction":{"count":315267,"self":571.2923648,"total":571.29239,"children":null},"AgentAct":{"count":315267,"self":3.1051964,"total":3.1051963,"children":null},"AgentInfo.ToProto":{"count":80,"self":0,"total":0.0049864,"children":{"GenerateSensorData":{"count":80,"self":0.00099669999999999989,"total":0.0049864,"children":{"RayPerceptionSensor.Perceive":{"count":80,"self":0.0039897,"total":0.0039897,"children":null}}}}}},"gauges":{"Tank.CumulativeReward":{"count":80,"max":1,"min":1,"runningAverage":1,"value":1,"weightedAverage":1}},"metadata":{"timer_format_version":"0.1.0","start_time_seconds":"1614532058","unity_version":"2019.4.18f1","command_line_arguments":"D:\\Unity\\2019.4.18f1\\Editor\\Unity.exe -projectpath D:\/Project\/UNITY\/ml-agents\/ML-tanks -useHub -hubIPC -cloudEnvironment production -hubSessionId 6926a880-795e-11eb-8b20-5718f046107e -accessToken RsoRu4LQuDe8JLNsMU1OiZfhtxhi139rTZ9vielB6hw00ef","communication_protocol_version":"1.3.0","com.unity.ml-agents_version":"1.7.2-preview","scene_name":"step1","end_time_seconds":"1614532923"}}
