+++ Assets/MLTank/Train_tankA/step1.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: c155a4a8f48a6b6478d3476798a0c5df
-PrefabImporter:
+guid: fdadb53c08388b74d818c055cad71689
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
