+++ ProjectSettings/TagManager.asset
@@ -7,6 +7,7 @@ TagManager:
   - Shell
   - Mine
   - block
+  - target
   layers:
   - Default
   - TransparentFX
